refactor: move model files inside aggregate package
