+++ src/application/session/entity.ts
