+++ src/application/session/utils.ts
