+++ src/presentation/commands/connection/session/createSession.ts
@@ -1,6 +1,6 @@
 import { Server, Socket } from 'socket.io'
 import { SessionCommandHandlers } from '../../../../application/commands/sessionCommandHandlers'
-import { Ack } from '../../../../application/message'
+import { Ack } from '../../../../application/session/message'
 import { CreateSessionCommand } from '../../../../application/commands/sessionCommands'
 
 /**
