+++ src/presentation/commands/videoCommands/videoCommands.ts
@@ -1,5 +1,5 @@
 import { Server } from 'socket.io'
-import { Ack } from '../../../application/message'
+import { Ack } from '../../../application/session/message'
 import { SessionCommandHandlers } from '../../../application/commands/sessionCommandHandlers'
 import { SessionNotifications } from '../../notifications/sessionNotifications'
 import { PlayVideoCommand, StopVideoCommand } from '../../../application/commands/videoCommands'
